requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==8.0.0
